+++ src/test/java/nu/aron/next/CommandInDirectoryTest.java
@@ -1,4 +1,4 @@
-package nu.aron.nextbuildnumber;
+package nu.aron.next;
 
 import org.junit.jupiter.api.Test;